move tests into load
